full_agentic_build/requirements.txt
openai==1.45.0
langchain==0.3.0
langchain-openai==0.2.2
chromadb==0.5.4
sentence-transformers==3.0.1
beautifulsoup4==4.12.3
streamlit==1.38.0
python-dotenv==1.0.1
tqdm==4.66.5
pandas==2.2.2
matplotlib==3.9.0
